Separate input component
